requirements.txt
h5py==2.8.0
imageio==2.19.3
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.1.0
progress==1.5
scipy==1.6.2
six==1.16.0
torch==1.13.1+cu116
tqdm==4.64.0
